refactor: collapse MinimalSearchField into SearchField
MinimalSearchField was an internal one-use helper; inline its markup and
styles into SearchField and delete it. Move the story to SearchField.
