+++ src/components/SearchField/SearchField.tsx
@@ -1,8 +1,11 @@
 import classNames from 'classnames'
-import { Search } from 'lucide-react'
+import { Search, X } from 'lucide-react'
+import { Input as ReactAriaInput } from 'react-aria-components/Input'
+import { SearchField as ReactAriaSearchField } from 'react-aria-components/SearchField'
+import { useTranslation } from 'react-i18next'
 
+import { Button } from '@ui/Button/Button'
 import { InputGroup } from '@ui/Input/InputGroup'
-import { MinimalSearchField } from '@ui/SearchField/MinimalSearchField'
 import { VStack } from '@ui/Stack/Stack'
 
 import './searchField.scss'
@@ -19,6 +22,7 @@ export type SearchFieldProps = {
 }
 
 export function SearchField({ slot = 'search', className, label, isDisabled, ...restProps }: SearchFieldProps) {
+  const { t } = useTranslation()
   const combinedClassName = classNames(CLASS_NAME, className)
 
   return (
@@ -26,12 +30,17 @@ export function SearchField({ slot = 'search', className, label, isDisabled, ...
       <VStack slot='icon' align='center' justify='center' className='Layer__SearchField__Icon' data-disabled={isDisabled || undefined}>
         <Search size={14} />
       </VStack>
-      <MinimalSearchField
+      <ReactAriaSearchField
         {...restProps}
-        placeholder={label}
         aria-label={label}
         isDisabled={isDisabled}
-      />
+        className='Layer__SearchField__Field'
+      >
+        <ReactAriaInput slot='input' placeholder={label} />
+        <Button slot='clear-button' inset icon variant='ghost' aria-label={t('ui:action.clear_search', 'Clear search')}>
+          <X size={18} />
+        </Button>
+      </ReactAriaSearchField>
     </InputGroup>
   )
 }
