+++ src/components/SearchField/searchField.scss
@@ -25,3 +25,57 @@
     }
   }
 }
+
+.Layer__SearchField__Field {
+  display: grid;
+  grid-template-areas: 'input clear-button';
+  grid-template-columns: minmax(0, 1fr) auto;
+  align-items: center;
+
+  border: none;
+
+  & > [slot='input'] {
+    grid-area: input;
+
+    padding-inline: var(--spacing-3xs);
+
+    border: none;
+    outline: none;
+
+    /*
+     * The background on mobile is grey by default.
+     */
+    background: none;
+
+    font-family: var(--font-family);
+    color: var(--color-base-1000);
+    caret-color: var(--color-base-1000);
+
+    &::-webkit-search-cancel-button,
+    &::-webkit-search-decoration {
+      appearance: none;
+    }
+  }
+
+  &[data-empty] > [slot='input'] {
+    grid-column: 1 / -1;
+  }
+
+  & > [slot='clear-button'] {
+    grid-area: clear-button;
+  }
+
+  &[data-empty] > [slot='clear-button'] {
+    display: none;
+  }
+
+  &[data-disabled] {
+    cursor: not-allowed;
+  }
+
+  &[data-disabled] > [slot='input'] {
+    cursor: not-allowed;
+    color: var(--color-base-500);
+    caret-color: var(--color-base-500);
+  }
+}
